==> scripts/1_cleaning_cory.Rmd <==
---
title: "Cory - data cleaning"
author: "Simon van Baal"
date: "26/10/2020"
output: html_document
  html_document:
    toc: TRUE
---

```{r libraries}

library(renv)
#renv::restore()

library(dplyr)
library(readr)
library(tidyr)
library(stringr)
library(here)
```

```{r loading data}
eligibility <- read_csv(here("data/proc/comb_proc_demo_cory.csv"))
preTestData <- read_csv(here("data/proc/comb_proc_pre-test_cory.csv"))
postTestData <- read_csv(here("data/proc/comb_proc_post-test_cory.csv"))
chatbotData <- read_csv(here("data/proc/comb_proc_chatbot_cory.csv"))

```


```{r eligibility cleaning}
eligibility <-
  eligibility |>
  select(StartDate,
         ParticipantID,
         Age,
         Sex,
         Vic,
         Childhood:FM) |>
  slice(-3) # remove duplicate entry

```


```{r pre-test data cleaning}
# Select relevant columns in a logical order
preTestData <- 
  preTestData |>
  select(StartDate,
         ParticipantID,
         Nationality,
         Concern_Rank_1:Study_Type)

# Transform to long data
# Before this step, you may look at the questions at the top row
# of the dataset.

preTestData <-
  preTestData |>
  filter(!is.na(ParticipantID)) |> 
  pivot_longer(
    `1_Attitude_1`:`3_Attitude_1`,
    names_to = "Question",
    values_transform = list(Attitude = as.numeric),
    values_to = "Attitude"
  ) |>
  rename(
    Concern_Rank = Concern_Rank_1,
    Necessary_Rank = Necessary_Rank_1,
    Effectiveness_Rank = Effectiveness_Rank_1,
    Intention = Intention_1,
    TestImportance = Test_1
  ) |>
  mutate(
    Concern_Rank = as.numeric(Concern_Rank),
    Necessary_Rank = as.numeric(Necessary_Rank),
    Effectiveness_Rank = as.numeric(Effectiveness_Rank),
    Intention = as.numeric(Intention),
    TestImportance = as.numeric(TestImportance)
  )

# Eliminate unnecessary text from the Question column and assign risk 
# levels to the Risk column

preTestData <-
  preTestData |>
  mutate(
    Question = as.numeric(str_extract(Question, "\\d{1,2}")),
    Risk = ifelse(
      Question %in% c("5", "7", "17", "23", "25", "29"),
      "High",
      ifelse(
        Question %in% c("1", "19", "21", "27"),
        "Low",
        ifelse(Question %in% c("3", "9", "11", "13", "15"),
               "Minimal",
               NA)
      )
    )
  )

# Generate list of participant in the three groups.
participantList <-
  tibble(ParticipantID = paste0("P", seq(1:60)),
         Condition = c(rep("Control", 17),
                       rep("Compassion", 17),
                       rep("Exponential Growth", 26)))

preTestData <-
  preTestData |>
  left_join(participantList, by = "ParticipantID")


# Create summaries per risk level and add it to the data set

attitudeAveragePreTest <-
  preTestData |>
  group_by(ParticipantID, Risk) |>
  summarise(Attitude_Average = mean(Attitude)) |>
  pivot_wider(names_from = "Risk", values_from = "Attitude_Average")

preTestData <-
  preTestData |>
  left_join(attitudeAveragePreTest, by = c("ParticipantID"))

rm(attitudeAveragePreTest, participantList)
```


```{r post-test data cleaning}
postTestData <- 
  postTestData |>
  select(StartDate,
         ParticipantID,
         Experience_1:Condition) |>
  rename(Interest = Experience_1,
         Intrigue = Experience_2,
         Focus = Experience_3,
         Inattention = Experience_4,
         Distraction = Experience_5,
         Enjoyment = Experience_6,
         Annoyance = Experience_7,
         Pleasure = Experience_8,
         Learning = Experience_9,
         Empathy = Experience_10,
         TestImportance = Test_1,
         Recommendation = Recommendation_1)

postTestData <-
  postTestData |>
  filter(!is.na(ParticipantID)) |> 
  pivot_longer(
    `2_Attitudes_1`:`20_Attitudes_1`,
    names_to = "Question",
    values_transform = list(Attitude = as.numeric),
    values_to = "Attitude"
  ) |>
  mutate_at(vars(Interest:Empathy, 
                 TestImportance, 
                 Recommendation),
            funs(as.numeric))

# Eliminate unnecessary text from the Question column and assign risk levels to the Risk column

postTestData <-
  postTestData |>
  mutate(
    Question = as.numeric(str_extract(Question, "\\d{1,2}")),
    Risk = ifelse(
    Question %in% c("2", "4", "8", "10", "12", "20", "22", "26", "28" , "30"),
    "High",
    ifelse(
      Question %in% c("6", "16", "18"),
      "Low",
      ifelse(Question %in% c("14", "24"),
             "Minimal",
             NA)
    )
  ))

attitudeAveragePostTest <- 
  postTestData |>
  group_by(ParticipantID, Risk) |>
  summarise(Attitude_Average = mean(Attitude)) |>
  pivot_wider(names_from = "Risk", values_from = "Attitude_Average")

postTestData <- 
  postTestData |>
  left_join(attitudeAveragePostTest, by = "ParticipantID")

rm(attitudeAveragePostTest)
```

```{r Cory Data Cleaning}
chatbotData <-
  chatbotData |>
  arrange(`Joined At (+11:00)`)

# create summary
chatbotSummary <-
  chatbotData |>
  group_by(testingLikelihoodSpread,
           testingLikelihoodEmpathy,
           testingLikelihoodControl) |>
  summarise(Count = n()) |>
  filter(
    !is.na(testingLikelihoodSpread) |
      !is.na(testingLikelihoodEmpathy) |
      !is.na(testingLikelihoodControl)
  ) |>
  mutate(
    testingLikelihood = factor(
      ifelse(
        !is.na(testingLikelihoodControl),
        testingLikelihoodControl,
        ifelse(
          !is.na(testingLikelihoodEmpathy),
          testingLikelihoodEmpathy,
          testingLikelihoodSpread
        )
      ),
      levels = c("Very Likely", "I don't know", "Very Unlikely")
    ),
    Condition = factor(
      ifelse(
        !is.na(testingLikelihoodControl),
        "Control",
        ifelse(
          !is.na(testingLikelihoodEmpathy),
          "Compassion",
          ifelse(!is.na(testingLikelihoodSpread),
                 "Exponential Growth")
        )
      ),
      levels = c("Exponential Growth", "Compassion", "Control")
    )
  )

# create simple data for analysis
chatbotData <-
  chatbotData |>
  group_by(testingLikelihoodSpread,
           testingLikelihoodEmpathy,
           testingLikelihoodControl) |>
  filter(
    !is.na(testingLikelihoodSpread) |
      !is.na(testingLikelihoodEmpathy) |
      !is.na(testingLikelihoodControl)
  ) |>
  mutate(
    testingLikelihood = factor(
      ifelse(
        !is.na(testingLikelihoodControl),
        testingLikelihoodControl,
        ifelse(
          !is.na(testingLikelihoodEmpathy),
          testingLikelihoodEmpathy,
          testingLikelihoodSpread
        )
      ),
      levels = c("Very Unlikely", "I don't know", "Very Likely")
    ),
    Condition = factor(
      ifelse(
        !is.na(testingLikelihoodControl),
        "Control",
        ifelse(
          !is.na(testingLikelihoodEmpathy),
          "Compassion",
          ifelse(!is.na(testingLikelihoodSpread),
                 "Exponential Growth")
        )
      ),
      levels = c("Control", "Compassion", "Exponential Growth")
    )
  ) |>
  ungroup() |>
  select(ParticipantID, testingLikelihood, Condition)


```


```{r ID linking}

# Link demographics to data from the chatbot
chatbotData <- 
  left_join(eligibility, chatbotData, by = "ParticipantID")

```


```{r Create pre-post comparison data set}

comparisonDataPre <-
  preTestData |>
  select(ParticipantID,
         Nationality,
         TestImportance,
         Risk,
         Question,
         Attitude,
         Study_Type,
         Condition) |>
  arrange(ParticipantID, Question)

# Make separate dataframe to add information to post-test data
nationalityData <- 
  comparisonDataPre |>
  group_by(ParticipantID, Nationality) |>
  summarise(n = n()) |>
  select(ParticipantID, Nationality)

comparisonDataPost <-
  postTestData |>
  select(ParticipantID,
         TestImportance,
         Risk,
         Question,
         Attitude,
         Study_Type,
         Condition) |>
  arrange(ParticipantID, Question)

# Add nationality column
comparisonDataPost <-
  comparisonDataPost |> 
  left_join(nationalityData, by = "ParticipantID")

# Bind the two datasets row-wise now that they have the same number of columns
comparisonData <-
  comparisonDataPre |>
  rbind(comparisonDataPost) |> 
  mutate(Timepoint = factor(Study_Type,
                             levels = c("Pre-Test",
                                        "Post-Test")),
         Condition = factor(Condition,
                            levels = c("Exponential Growth",
                                       "Compassion",
                                       "Control")),
         Attitude = Attitude - 50,
         Question = factor(Question)) |>
  left_join(eligibility, by = "ParticipantID")

rm(nationalityData)
```

```{r Enter additional demographic data manually}
# We received some of the missing demographic data from participants directly
# We enter this manually here.

comparisonData <- 
  comparisonData |>
  mutate(Nationality = ifelse(ParticipantID == "P16", 
                               "Australia", 
                               ifelse(ParticipantID == "P37", 
                                      "Australia",
                                      ifelse(ParticipantID == "P38", 
                                             "Australia", 
                                             Nationality))),
         Age = ifelse(ParticipantID == "P16", 
                               29, 
                               ifelse(ParticipantID == "P37", 
                                      27,
                                      ifelse(ParticipantID == "P38", 
                                             20, 
                                             Age))))

```



```{r finish up}
# write the cleaned data
write_csv(comparisonData, here("data/clean/clean_surveys_cory.csv"))
write_csv(chatbotData, here("data/clean/clean_chatbot_cory.csv"))

# write the postTestData for the cory experience figure in the figures.rmd file
write_csv(postTestData, here("data/clean/clean_post-test_cory.csv"))

# clear environment
rm(list=ls())

```
