==> scripts/3_figures_cory.Rmd <==
---
title: "Cory - figures"
author: "Simon van Baal"
date: "17/03/2022"
output: html_document
---

```{r setup}

# Please run this script after 'analysis_cory.Rmd'.

library(ggplot2)
library(ggpubr)
library(kableExtra)

```

```{r load data}
# Please ensure you have run previous scripts.

postTestData <-
  read_csv(here("data/clean/clean_post-test_cory.csv"))

surveyData <- 
  read_csv(here("data/clean/clean_surveys_cory.csv"))

chatbotData <-
  read_csv(here("data/clean/clean_chatbot_cory.csv"))

```


# Tables

```{r Table sample demographics}

# Create a place to store the table
if (!dir.exists(here("output/tables"))) {
  dir.create(here("output/tables"))
}

# Reload data to deal with centering of age
chatbotData <-
  read_csv(here("data/clean/clean_chatbot_cory.csv"))

# Here we provide a table for demographic information of the sample.

demographics <-
  chatbotData |>
  filter(!is.na(Age)) |>
  mutate(coiGroup = factor( # country of origin group
    ifelse(
      Childhood != "Australia" &
        Childhood != "Viet Nam",
      "Other Migrants",
      Childhood
    ), levels = c(
      "Australia",
      "Viet Nam", 
      "Other Migrants"
    )
  )) |>
  group_by(ParticipantID, coiGroup, Age, Sex) |>
  summarise(n = n()) |>
  group_by(coiGroup, Sex) |>
  summarise(
    `M (n)` = round(mean(Age, na.rm = T), 1),
    SD = as.character(round(sd(Age, na.rm = T), 1)),
    n = n()) |>
  mutate(`M (n)` = paste0(`M (n)`, " (", n, ")")) |>
  arrange(coiGroup) |> # change order
  ungroup() |>
  select(Sex, `M (n)`, SD) |>
  rename(` ` = Sex)

kable(demographics, caption = "Sample demographics.") |>
  kable_classic(
    full_width = F,
    font_size = 14
  ) |>
  column_spec(1:3, color = "black") |>
  pack_rows("Australia", 1, 2) |>
  pack_rows("Viet Nam", 3, 3) |>
  pack_rows("Other Migrants", 4, 5) 


```


```{r table-survey-demographics}

# here we provide a table for demographic information of the sample used for
# the perceived importance of testing (pre/post chatbot) - partitioned by group.
# Some participants did not complete both surveys, So there is missing data.

# ----------------------------------- Testing importance table section
demographicsTestImp <-
  chatbotData |>
  filter(ParticipantID %in% testImportanceData$ParticipantID,
         !is.na(Age), 
         !is.na(Childhood)) |>
  mutate(coiGroup = factor(
    ifelse(
      Childhood != "Australia" &
        Childhood != "Viet Nam",
      "Other Migrants",
      Childhood
    )
  )) |>
  group_by(ParticipantID, coiGroup, Condition, Age, Sex) |>
  summarise(n = n()) 

list <- 
  surveyData |>
  group_by(ParticipantID, Condition) |>
  summarise(n = n()) |>
  select(ParticipantID, Condition)

demographicsTestImp <- 
  demographicsTestImp |> 
  left_join(list, by = "ParticipantID") |>
  rename(Condition = Condition.y)

sexInfoTestImp <- 
  demographicsTestImp |>
  group_by(Condition, Sex) |>
  summarise(n = n())
  
demographicsTestImp <-
  demographicsTestImp |>
  group_by(Condition, coiGroup) |>
  summarise(
    `M (n)` = round(mean(Age, na.rm = T), 1),
    SD = as.character(round(sd(Age, na.rm = T), 1)),
    n = n()) |>
  mutate(`M (n)` = paste0(`M (n)`, " (", n, ")")) |>
  select(Condition, coiGroup, `M (n)`, SD)

# Create dfs for the different conditions to make table
tableDf1 <-
  demographicsTestImp |>
  filter(Condition == "Compassion") |>
  rename(` ` = coiGroup) |>
  ungroup() |>
  select(` `, `M (n)`, SD) |>
  add_row(` ` = "Viet Nam", `M (n)` = "-" , SD = "-") |>
  mutate(SD = ifelse(is.na(SD), "-", SD))

tableDf2 <-
  demographicsTestImp |>
  filter(Condition == "Exponential Growth") |>
  ungroup() |>
  select(`M (n)`, SD) |>
  mutate(SD = ifelse(is.na(SD), "-", SD))

tableDf3 <-
  demographicsTestImp |>
  filter(Condition == "Control") |>
  ungroup() |>
  select(`M (n)`, SD) |>
  add_row(`M (n)` = "-", SD = "-") |>
  mutate(SD = ifelse(is.na(SD), "-", SD))


# ----------------------------------- Testing likelihood table section

# Many participants did not answer Cory COVID-bot's question on how likely 
# they were to get tested if they experienced symptoms; it was voluntary.
# Here we describe the sample for the testing likelihood data.

# Do this over again with only people who replied to Cory's question
demographicsTestLik <-
  chatbotData |>
  filter(!is.na(Age), !is.na(Childhood),
         !is.na(testingLikelihood)) |>
  mutate(coiGroup = factor(
    ifelse(
      Childhood != "Australia" &
        Childhood != "Viet Nam",
      "Other Migrants",
      Childhood
    )
  )) |>
  group_by(ParticipantID, coiGroup, Condition, Age, Sex) |>
  summarise(n = n())

sexInfoTestLik <-
  demographicsTestLik |>
  group_by(Condition, Sex) |>
  summarise(n = n())

demographicsTestLik <-
  demographicsTestLik |>
  group_by(Condition, coiGroup) |>
  summarise(
    `M (n)` = round(mean(Age, na.rm = T), 1),
    SD = as.character(round(sd(Age, na.rm = T), 1)),
    n = n()) |>
  mutate(`M (n)` = paste0(`M (n)`, " (", n, ")")) |>
  select(Condition, coiGroup, `M (n)`, SD)

tableDf4 <-
  demographicsTestLik |>
  filter(Condition == "Compassion") |>
  rename(` ` = coiGroup) |>
  ungroup() |>
  select(`M (n)`, SD) |>
  add_row(`M (n)` = "-" , SD = "-") |>
  mutate(SD = ifelse(is.na(SD), "-", SD))

tableDf5 <-
  demographicsTestLik |>
  filter(Condition == "Exponential Growth") |>
  ungroup() |>
  select(`M (n)`, SD) |>
  mutate(SD = ifelse(is.na(SD), "-", SD))

tableDf6 <-
  demographicsTestLik |>
  filter(Condition == "Control") |>
  ungroup() |>
  select(`M (n)`, SD) |>
  add_row(`M (n)` = "-", SD = "-") |>
  mutate(SD = ifelse(is.na(SD), "-", SD))

# ---------------------------- Sex information for table

sexInfoTestImp <-
  sexInfoTestImp |>
  pivot_wider(id_cols = Condition,
              names_from = Sex,
              values_from = n) |>
  mutate(`F/M` = paste0(Female, "/", Male))

sexInfoTestLik <-
  sexInfoTestLik |>
pivot_wider(id_cols = Condition,
              names_from = Sex,
              values_from = n) |>
  mutate(`F/M` = paste0(Female, "/", Male))

sexInfoList <-
  c(sexInfoTestImp$`F/M`, sexInfoTestLik$`F/M`)


```


```{r store-table}
tableDfComplete <- 
  cbind(tableDf1, tableDf2, tableDf3,
        tableDf4, tableDf5, tableDf6)

tableDfComplete |>
  kable(caption = "Table 1. Sample demographics by condition.") |>
  kable_classic(html_font = "Cambria", full_width = F) |>
  column_spec(1:13, color = "black") |>
  add_header_above(c(
    "Female/Male" = 1,
    "7/6" = 2,
    "8/6" = 2,
    "17/3" = 2,
    "7/2" = 2,
    "4/3" = 2,
    "6/2" = 2
  )) |>
  add_header_above(c(
    " " = 1,
    "Compassion" = 2,
    "Exponential" = 2,
    "Control" = 2,
    "Compassion" = 2,
    "Exponential" = 2,
    "Control" = 2
  )) |>
  add_header_above(c(
    " " = 1,
    "Perceived Importance of Testing" = 6,
    "Likelihood of Getting Tested" = 6
  )) |> 
  save_kable(file = 
               here("output/tables/table_demographics_by-condition.html"))


rm(tableDf1,
   tableDf2,
   tableDf3,
   tableDf4,
   tableDf5,
   tableDf6,
   list)

```

# Figures

## Figures of Chatbot Evaluation

```{r Chatbot Evaluation}
if (!dir.exists(here("output/figures"))) {
  dir.create(here("output/figures"))
}

DCBIavg <-
  postTestData |>
  summarise(
    Interest_mean = mean(Interest),
    Interest_sd = sd(Interest),
    Intrigue_mean = mean(Intrigue),
    Intrigue_sd = sd(Intrigue),
    Focus_mean = mean(Focus),
    Focus_sd = sd(Focus),
    Inattention_mean = mean(Inattention),
    Inattention_sd = sd(Inattention),
    Distraction_mean = mean(Distraction),
    Distraction_sd = sd(Distraction),
    Enjoyment_mean = mean(Enjoyment),
    Enjoyment_sd = sd(Enjoyment),
    Annoyance_mean = mean(Annoyance),
    Annoyance_sd = sd(Annoyance),
    Pleasure_mean = mean(Pleasure),
    Pleasure_sd = sd(Pleasure),
    Learning_mean = mean(Learning),
    Learning_sd = sd(Learning)
  ) |>
  pivot_longer(
    Interest_mean:Learning_sd,
    names_to = c("Metric", "Type"),
    names_sep = "_",
    values_to = c("Response")
  ) |>
  pivot_wider(id_cols = "Metric",
              names_from = "Type",
              values_from = "Response")

## DCBI plot

plotDCBI <- 
  ggplot(DCBIavg,
       aes(
         x = reorder(Metric, desc(mean)),
         y = mean,
         col = mean
       )) +
  geom_point() +
  geom_errorbar(aes(ymin = mean - sd, ymax = mean + sd),
                width = .2) +
  labs(y = "Likert Scale Response", x = "Metric",
       col = "Response") +
  scale_y_continuous(breaks = seq(1, 7, 1), limits = c(.94, 7.01)) +
  scale_colour_viridis_c(begin = .1, end = .65) +
  theme_light() +
  theme(legend.position = "none")
ggsave(
    here("output/figures/plot_experience_cory.png"),
    height = 4,
    width = 6
  )

rm(DCBIavg)

```

## Figures of Testing Attitudes

```{r Testing-likelihood-figs}

# Testing likelihood figure

plotTestingLikelihood <-
  ggplot(
    chatbotData |>
      filter(!is.na(testingLikelihood)) |>
      mutate(
        testingLikelihood = factor(
          ifelse(
            testingLikelihood == "Very Likely",
            "Very Likely",
            ifelse(testingLikelihood == "I don't know",
                   "Unsure",
                   "Very Unlikely")
          ),
          levels = c("Very Likely",
                     "Unsure",
                     "Very Unlikely")
        ),
        Condition = factor(
          Condition,
          levels = c("Exponential Growth",
                     "Compassion",
                     "Control")
        )
      ),
    aes(group = Condition,
        x = testingLikelihood)
  ) +
  geom_bar(aes(y = ..prop.., fill = Condition),
           width = .5,
           position = position_dodge2()) +
  labs(x = "Likelihood of Getting tested",
       y = "Percentage of Answers") +
  theme_light() +
  theme(legend.position = "bottom") +
  scale_fill_viridis_d(begin = .2, end = .8) +
  scale_y_continuous(labels = scales::percent)
ggsave(here("output/figures/plot_test-likelihood.png"),
            width = 6, height = 4)


```


```{r Testing importance attitudes figures}
# Run simple model for figure

lmTestImportance <- 
  lm(testImportanceDifference ~
       Condition,
     data = testImportanceData)

plotTestImportance <- 
  afex_plot(lmTestImportance, ~Condition,
          mapping = "colour",
          data_geom = geom_violin) +
  geom_hline(aes(yintercept = 0),
             alpha = .5,
             size = .2) +
  labs(y = expression("Positive Change in Perceived Testing Importance " %->% "")) +
  scale_color_viridis_d(begin = .2, end = .65) +
  theme_light() +
  theme(legend.position = "none")

ggsave(here("output/figures/plot_test-importance.png"),
       height = 4, 
       width = 6)

```


## Figures of Attitudes toward Leaving Home

```{r plots for attitudes to leaving home attitudes}
#=================================================== Density plots for attitudes
 
surveyData <- 
  surveyData |>
  mutate(TimepointNumeric = ifelse(Timepoint == "Pre-Test",0,1))

plotHighRiskAttitude <-
  ggplot(surveyData |>
           filter(Risk == "High"),
         aes(x = Attitude, fill = Timepoint)) +
  geom_density(alpha = .4) +
  geom_boxplot(
    aes(x = Attitude, 
        y = TimepointNumeric * .0015 + 0.0315),
    alpha = .4,
    colour = "black",
    notch = F,
    coef = 1,
    width = .001,
  ) +
  labs(title = "High Risk", 
       x = "Certainty", 
       y = "Frequency") +
  scale_fill_viridis_d(begin = .2, end = .7) +
  theme_light() +
  theme(axis.title.x = element_blank()) +
  guides(fill = guide_legend(reverse = T))

plotLowRiskAttitude <-
  ggplot(surveyData |>
           filter(Risk == "Low"),
         aes(x = Attitude, fill = Timepoint)) +
  geom_density(alpha = .4) +
  geom_boxplot(aes(x = Attitude, 
                   y = TimepointNumeric * .0015 + 0.0315),
               alpha = .4, 
               colour = "black", 
               notch = F, 
               coef = 1,
               width = .001) +
  labs(title = "Low Risk", 
       x = "Certainty", 
       y = "Frequency") +
  scale_fill_viridis_d(begin = .2, end = .7) +
  theme_light() +
  theme(axis.title = element_blank()) +
  guides(fill = guide_legend(reverse = T)) 

plotMinimalRiskAttitude <-
  ggplot(surveyData |>
           filter(Risk == "Minimal"),
         aes(x = Attitude, fill = Timepoint)) +
  geom_density(alpha = .4) +
  geom_boxplot(aes(x = Attitude, y = TimepointNumeric * .0015 + 0.0315),
               alpha = .4, 
               colour = "black", 
               notch = F, 
               coef = 1,
               width = .001) +
  labs(title = "Minimal Risk",
       x = "Certainty", 
       y = "Frequency",
       fill = "Timepoint") +
  scale_fill_viridis_d(begin = .2, end = .7) +
  theme_light() +
  theme(axis.title = element_blank()) +
  guides(fill = guide_legend(reverse = T))
 

attitudePlots <- 
  ggarrange(plotHighRiskAttitude,
          plotLowRiskAttitude,
          plotMinimalRiskAttitude,
          ncol = 3, nrow = 1,
          common.legend = T,
          legend = "top")

attitudePlots <- 
  annotate_figure(attitudePlots, 
                bottom = text_grob("Permissiveness of Going Out"))
ggsave(here("output/figures/plot_certainty-pre-post.png"),
         height = 4, width = 8)

```
